+++ leads_vec/__init__.py
@@ -4,3 +4,4 @@ if not _find_spec("pynput"):
     raise ImportError("Please install `pynput` to run this module\n>>>pip install pynput")
 
 from leads_vec.run import *
+from leads_vec.__entry__ import __entry__
